+++ src/containerapp/azext_containerapp/commands.py
@@ -223,3 +223,8 @@ def load_command_table(self, args):
         g.custom_command('create', 'create_session_pool', supports_no_wait=True)
         g.custom_command('update', 'update_session_pool', supports_no_wait=True)
         g.custom_command('delete', 'delete_session_pool', confirmation=True, supports_no_wait=True)
+
+    with self.command_group('containerapp java logger', is_preview=True) as g:
+        g.custom_command('set', 'create_or_update_java_logger', supports_no_wait=True)
+        g.custom_command('delete', 'delete_java_logger', supports_no_wait=True)
+        g.custom_show_command('show', 'show_java_logger')